Move logging to new _gpi_embed module
